Add files via upload
A landing page is known as the lead capture page, static page, or destination page which is designed for a specific task and appears because of an online advertisement or a search engine optimization result. When the user arrives on a landing page it is free from all distractions like the home page and mainly aims to capture a visitor’s information through that form and make them feel comfortable that the website has the solution to their problems. In this article, we are going to see how to create a Landing page using HTML CSS, and JavaScript.

Approach
Create the basic layout of the landing page using HTML that includes a basic nav bar, a basic homepage a body section, and a signup page.
Style the layout created by HTML using CSS and link the CSS file to the HTML file.
Create an index.js file and write the JavaScript code to display the nav bar in a sliding menu when the website is viewed on small screens